Exit on error

+++ workloads/kube-burner/common.sh
@@ -29,6 +29,9 @@ collect_pprof() {
 
 deploy_operator() {
   deploy_benchmark_operator ${OPERATOR_REPO} ${OPERATOR_BRANCH}
+  if [[ $? != 0 ]]; then
+    exit 1
+  fi
   rm -rf benchmark-operator
   git clone --single-branch --branch ${OPERATOR_BRANCH} ${OPERATOR_REPO} --depth 1
   kubectl apply -f benchmark-operator/resources/backpack_role.yaml
@@ -58,11 +61,12 @@ run_workload() {
   local TMPCR=$(mktemp)
   envsubst < $1 > ${TMPCR}
   run_benchmark ${TMPCR} $((JOB_TIMEOUT + 600))
-  kubectl delete -f ${TMPCR}
+  local rc=$?
   if [[ ${TEST_CLEANUP} == "true" ]]; then
     log "Cleaning up benchmark"
     kubectl delete -f ${TMPCR}
   fi
+  return ${rc}
 }
 
 label_nodes() {
